skilllite-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skilllite"
version = "0.1.0"
description = "A lightweight Skills execution engine with LLM integration for LLM agents"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "SkillLite Team", email = "skilllite@example.com"}
]
keywords = ["agent", "skills", "llm", "claude", "openai", "sandbox", "tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
mcp = ["mcp>=1.0.0"]
all = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "mcp>=1.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/skilllite/skilllite"
Documentation = "https://github.com/skilllite/skilllite#readme"
Repository = "https://github.com/skilllite/skilllite"
Issues = "https://github.com/skilllite/skilllite/issues"

[project.scripts]
skilllite = "skilllite.cli:main"
skilllite-mcp = "skilllite.mcp:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["skilllite*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
